EVM: Removes knowledge of TobinTax
EVM should be unaware of tobin tax. From the EVM perspective there is
just the Context.Transfer() function.

This commit, moves evm.TobinTransfer to replace context.Transfer
function. In so, it removes unnecessary input and output parameters.
