+++ core/vm/contracts.go
@@ -634,7 +634,7 @@ func (c *transfer) Run(input []byte, caller common.Address, evm *EVM, gas uint64
 			return nil, gas, ErrInsufficientBalance
 		}
 
-		gas, err = evm.TobinTransfer(evm.StateDB, from, to, gas, value)
+		evm.Context.Transfer(evm, from, to, value)
 	}
 
 	return input, gas, err
